* pixelX, pixelY can be float
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@1018 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ include/camera_info.h
@@ -21,8 +21,8 @@ struct chip_info
   int height;
   int binning_vertical;
   int binning_horizontal;
-  int pixelX;
-  int pixelY;
+  float pixelX;
+  float pixelY;
   int image_type;		/* see above for types */
   float gain;			/* in 0.01 e / ADU */
 };
